WIP: Fix hot standby
Can't trim undo logs on the hot standby, because it writes WAL and
buffers. Need to wait for those changes to be replicated from the
master. Instead, just return fast from zsundo_trim() with the oldest
undo log pointer from the metapage, and do the visibility checks the
'hard' way.