Compressed toast datums
Previously datum greater than MaxZedStoreDatumSize is stored in seperate toast
pages uncompressed, this can be waste of space. This patch adds compression to
the toast datum and depending on the compressed size, store them either in line
or in separate toast pages.

If the compressed datum is small enough (not greater than
MaxZedStoreDatumSize), then store it inline, otherwise, store it in separate
toast pages.